src/requirements.txt
cffi==1.17.1
cryptography==44.0.1
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
